requirements.txt
attrs >= 21.3.0
PyJWT >= 2.6.0, < 3
requests >= 2.28.0, < 3
cryptography >= 40.0.0, < 42
pyOpenSSL >= 23.1.1, < 24
asn1==2.7.0
cattrs==23.1.2
